+++ libs/mllegos/src/thds/mllegos/sklegos/feature_extraction/__init__.py
@@ -1,3 +1,17 @@
-__all__ = ["OptimizedPrefixEncoding"]
+__all__ = [
+    "DenseToSparseNamedTransformer",
+    "OptimizedPrefixEncoding",
+    "SparseCountFeatureNormalized",
+    "SparseCountFeatures",
+    "to_dense",
+    "to_sparse",
+]
 
 from .prefix_coding import OptimizedPrefixEncoding
+from .sparse import (
+    DenseToSparseNamedTransformer,
+    SparseCountFeatureNormalized,
+    SparseCountFeatures,
+    to_dense,
+    to_sparse,
+)